S12P31E103-169 [BE] refactor: 라운드 종료 시 로직 수정

See merge request s12-final/S12P31E103!254

+++ Backend/src/main/java/kr/codenova/backend/multi/dto/broadcast/GameResultBroadcast.java
@@ -21,7 +21,7 @@ public class GameResultBroadcast {
     @Builder
     public static class UserResultStatus {
         private String nickname;
-        private Double averageScore;
+        private Integer averageScore;
         private int rank;
     }
 }
